Move findnz to SparseArrays module
Its definition (return stored entries) only makes sense in the context of sparse arrays.
